# src/main/resources/static/swagger-ui/openapi.yaml
openapi: 3.0.0
info:
  title: API de Residencias
  version: 1.0.0
  description: API para gestionar residencias, con operaciones de registro y consulta.
paths:
...
  /resi/add:
    post:
      tags:
        - Residencia
      summary: Registrar una nueva residencia
      description: Crea una nueva residencia validando campos obligatorios, formato de correo y unicidad.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Residencia.Dto'
      responses:
        '201':
          description: Residencia registrada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Residencia.ResponseDto'
        '400':
          description: Datos del cuerpo nulos/vacios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: Correo Invalido, tiene que ser un correo con extructura correcta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CORREO_INVALIDO:
                  value:
                    mensaje: Email invalid
                    codigo: 1002
                    status: 406
                    timestamp: 2025-04-13T21:00:0
        '409':
          description: Email o nombre que han sido ingreado en el cuerpo ya existen en otra residencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                CORREO_DUPLICADO:
                  value:
                    mensaje: Email ya existente
                    codigo: 1003
                    status: 409
                    timestamp: 2025-04-13T21:00:00
                NOMBRE_DUPLICADO:
                  value:
                    mensaje: Nombre ya existente
                    codigo: 1004
                    status: 409
                    timestamp: 2025-04-13T21:00:00
  /resi/{idResidencia}/get:
    get:
      tags:
        - Residencia
      summary: Busca la residencia por id ingresado en el path
      description: Devuelve un dto Response con la residencia encontrada.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Residencia encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Residencia.ResponseDto'
        '400':
          description: Datos nulos/vacios de la variable de la url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Residencia invalida, no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                  value:
                    mensaje: Residencia invalida
                    codigo: 1006
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
...
  /resi/resident/add:
    post:
      tags:
        - Residente
      summary: Registrar un nuevo residente.
      description: Crea un residente en una residencia existente en la base de datos, con ciertos campos.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Residente.Dto'
      responses:
        '201':
          description: Residente registrado con existo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Residente.ResponseDto'
        '400':
          description: Datos del cuerpo nulos/vacios o de la variable del path idResidencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                  value:
                    mensaje: Residencia invalida
                    codigo: 1006
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: Fecha invalida o docuemnto de identidad invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                FECHA_INVALIDO:
                  value:
                    mensaje: Fecha invalida
                    codigo: 1005
                    status: 406
                    timestamp: 2025-04-13T21:00:00
                DOCUMENTO_INVALIDO:
                  value:
                    mensaje: Documento de identidad no tiene el formato correcto
                    codigo: 1027
                    status: 406
                    timestamp: 2025-04-13T21:00:00
        '409':
          description: El documento de identidad ya existe en la residencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                DOCUMENTO_DUPLICADO:
                  value:
                    mensaje: Documento de identidad ya existente
                    codigo: 1028
                    status: 409
                    timestamp: 2025-04-13T21:00:00
  /resi/resident/{idResidente}/get:
    get:
      tags:
        - Residente
      summary: Busca el residente por id ingresado en el path
      description: Devuelve un dto Response con el residente encontrada.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidente
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Residente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Residente.ResponseDto'
        '400':
          description: Datos nulos/vacios de la variable de la url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Residente invalido, no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENTE_INVALIDO:
                  value:
                    mensaje: Residente invalida
                    codigo: 1009
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
  /resi/resident/getAll:
    get:
      tags:
        - Residente
      summary: Obtener residentes
      description: Devuelve una lista de residentes.  Y filtrado por los sigueintes parametros. Hay que tener en cuenta sigue una proridad, documento de identidad, fecha de nacimiento y mes/año
      security:
        - bearerAuth: []
      parameters:
        - name: documentoIdentidad
          in: query
          required: false
          schema:
            type: string
        - name: fechaNacimiento
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          required: false
          schema:
              type: integer
              format: int64
        - name: maxAge
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: minAge
          in: query
          required: false
          schema:
              type: integer
              format: int64
        - name: idJuego
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: idEventoSalida
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Residente(s) econtrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Residente.ResponseDto'
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                value:
                  mensaje: Residencia invalida
                  codigo: 1006
                  status: 404
                  timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: Rango de edad invalido, el minimo no puede ser mayor que el maximo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RANGO_EDAD_INVALIDO:
                  value:
                    mensaje: Rango de edad invalido
                    codigo: 1029
                    status: 406
                    timestamp: 2025-04-13T21:00:00
  /resi/resident/getAll/bajas:
    get:
      tags:
        - Residente
      summary: Obtener residentes en baja
      description: Devuelve una lista de residentes en baja.
      security:
        - bearerAuth: []
      parameters:
        - name: documentoIdentidad
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Residente(s) econtrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Residente.ResponseDto'
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                value:
                  mensaje: Residencia invalida
                  codigo: 1006
                  status: 404
                  timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
...
  /resi/resident/{idResidente}/baja:
    patch:
      tags:
        - Residente
      summary: Dar de baja a un residente
      description: Cambia el estado de un residente a baja.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidente
          in: query
          required: true
          schema:
              type: integer
              format: int64
      responses:
        '200':
          description: Residente dado de baja
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Residente.ResponseDto'
        '400':
          description: Campos nulos/vacios de las variable path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                RESIDENCIA_INVALIDO:
                  value:
                    mensaje: Residencia invalida
                    codigo: 1006
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                RESIDENTE_INVALIDO:
                  value:
                    mensaje: Residente invalida
                    codigo: 1009
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: Residente ya dado de baja
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENTE_BAJA:
                  value:
                    mensaje: Residente ya dado de baja
                    codigo: 1028
                    status: 406
                    timestamp: 2025-04-13T21:00:00
  /resi/resident/{idResidente}/update:
    patch:
      tags:
        - Residente
      summary: Actualiza los datos de un residente
      description: Campos como nombre, apellido, documentoIdentidad, fecha de nacimiento.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidente
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Residente.Dto'
      responses:
        '200':
          description: Residente actualizado
          content:
              application/json:
                schema:
                    $ref: '#/components/schemas/Residente.ResponseDto'
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0

        '406':
          description: Documento de identidad invalido o fecha invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                FECHA_INVALIDO:
                  value:
                    mensaje: Fecha invalida
                    codigo: 1005
                    status: 406
                    timestamp: 2025-04-13T21:00:00
                DOCUMENTO_INVALIDO:
                  value:
                    mensaje: Documento de identidad no tiene el formato correcto
                    codigo: 1027
                    status: 406
                    timestamp: 2025-04-13T21:00:00
                RESIDENTE_BAJA:
                  value:
                    mensaje: Residente ya dado de baja
                    codigo: 1028
                    status: 406
                    timestamp: 2025-04-13T21:00:00
        '409':
          description: El documento de identidad ya existe en la residencia
          content:
              application/json:
                schema:
                    $ref: '#/components/schemas/Error.ResponseDto'
                example:
                    DOCUMENTO_DUPLICADO:
                    value:
                        mensaje: Documento de identidad ya existente
                        codigo: 1028
                        status: 409
                        timestamp: 2025-04-13T21:00:00
...
components:
  schemas:
...
    Residencia.Dto:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          example: Residencia La Paz
        email:
          type: string
          example: contacto@resilapaz.com
    Residencia.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Residencia La Paz
        email:
          type: string
          example: contacto@resilapaz.com
...
    Residente.Dto:
      type: object
      required:
        - nombre
        - apellido
        - fechaNacimiento
        - idResidencia
        - documentoIdentidad
      properties:
        nombre:
          type: string
          example: Carlos
        apellido:
          type: string
          example: Pérez
        fechaNacimiento:
          type: string
          format: date
          example: 1945-06-15
        idResidencia:
          type: integer
          format: int64
          example: 1
        documentoIdentidad:
            type: string
            example: Y8440460X
    Residente.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        nombre:
          type: string
          example: Carlos
        apellido:
          type: string
          example: Pérez
        fechaNacimiento:
          type: string
          format: date
          example: 1945-06-15
        idResidencia:
          type: integer
          format: int64
          example: 1
        documentoIdentidad:
          type: string
          example: Y8440460X
        baja:
          type: boolean
          example: false
...
    Error.ResponseDto:
      type: object
      properties:
        mensaje:
          type: string
        codigo:
          type: integer
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        mensaje: Faltan campos obligatorios
        codigo: 1001
        status: 400
        timestamp: 2025-04-13T21:00:00
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT